config restructuring
