+++ comms/trending/db/clickhouse.go
@@ -1,9 +1,24 @@
 package db
 
-type ClickHouseDB struct{}
+import (
+	"comms.audius.co/trending/config"
+	"github.com/ClickHouse/clickhouse-go/v2"
+)
 
-func NewClickHouseDB() *ClickHouseDB {
-	return &ClickHouseDB{}
+type ClickHouseDB struct {
+	CHDB clickhouse.Conn
+}
+
+func NewClickHouseDB(conf config.Config) (*ClickHouseDB, error) {
+	conn, err := clickhouse.Open(&clickhouse.Options{})
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &ClickHouseDB{
+		CHDB: conn,
+	}, nil
 }
 
 func (ch ClickHouseDB) CheckHealth() bool {
