+++ comms/trending/trending_main.go
@@ -2,11 +2,18 @@ package trending
 
 import (
 	"log"
+
+	"comms.audius.co/trending/config"
 )
 
 func TrendingMain() {
-	conf := DefaultConfig()
-	ts := NewTrendingServer()
-	err := ts.Run(*conf)
+	conf := config.DefaultConfig()
+
+	ts, err := NewTrendingServer(*conf)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	err = ts.Run()
 	log.Fatal(err)
 }
