+++ comms/trending/trendingserver.go
@@ -3,6 +3,7 @@ package trending
 import (
 	"net/http"
 
+	"comms.audius.co/trending/config"
 	"comms.audius.co/trending/db"
 	"github.com/labstack/echo/v4"
 )
@@ -10,29 +11,35 @@ import (
 type TrendingServer struct {
 	Database  db.DB
 	WebServer *echo.Echo
+	Conf      config.Config
 }
 
-func NewTrendingServer() *TrendingServer {
+func NewTrendingServer(conf config.Config) (*TrendingServer, error) {
 	wsv := echo.New()
 	wsv.Debug = true
 
 	// can swap with any DB type
-	db := db.NewClickHouseDB()
+	db, err := db.NewClickHouseDB(conf)
+
+	if err != nil {
+		return nil, err
+	}
 
 	return &TrendingServer{
 		WebServer: wsv,
 		Database:  db,
-	}
+		Conf:      conf,
+	}, nil
 }
 
-func (ts *TrendingServer) Run(conf Config) error {
+func (ts *TrendingServer) Run() error {
 	wsv := ts.WebServer
 
 	wsv.GET("/tracks/trending", ts.GetTrendingTracks)
 	wsv.GET("/playlists/trending", ts.GetTrendingPlaylists)
 	wsv.GET("/trending/health", ts.GetHealth)
 
-	return wsv.Start(conf.WebServerPort)
+	return wsv.Start(ts.Conf.WebServerPort)
 }
 
 func (ts *TrendingServer) GetTrendingTracks(c echo.Context) error {
